apps/odds-radar-chromium/pyproject.toml
[project]
name = "sheva-odds-radar"
version = "0.1.0"
description = "Radar de odds Bet365 com Playwright Chromium + Telegram"
requires-python = ">=3.11"

dependencies = [
    "playwright>=1.49.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[project.scripts]
radar = "src.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"
